refactor(types): colocate ColumnAlignment enum with other global option enums
Moves ColumnAlignment and ColumnAlignmentOptions from
src/components/options/columnstyles/types.ts to src/types.ts, next to
ColumnStyleColoring/ColorModeOptions, DatatablePagingType/
DatatablePagingOptions, and AggregationType/AggregationOptions. All other
enum+options pairs live in src/types.ts; keeping this one isolated in the
columnstyles dir broke the pattern.

Consumers now import from 'types' directly. Zero behavior change; 47 unit
tests still pass.
