+++ src/migrations.ts
@@ -7,6 +7,7 @@ import {
 
 import {
   AggregationOptions,
+  ColumnAlignment,
   ColumnAliasField,
   ColumnSorting,
   ColumnSortingOptions,
@@ -16,7 +17,7 @@ import {
   TransformationOptions,
 } from './types';
 import { Threshold } from 'components/options/thresholds/types';
-import { ColumnAlignment, ColumnStyleDate, ColumnStyleHidden, ColumnStyleItemType, ColumnStyleMetric, ColumnStyles, ColumnStyleString } from 'components/options/columnstyles/types';
+import { ColumnStyleDate, ColumnStyleHidden, ColumnStyleItemType, ColumnStyleMetric, ColumnStyles, ColumnStyleString } from 'components/options/columnstyles/types';
 //import { DEFAULT_CRITICAL_COLOR_RGBA, DEFAULT_OK_COLOR_RGBA, DEFAULT_WARNING_COLOR_RGBA } from 'components/options/defaults';
 
 interface AngularDatatableOptions {
